Move Config,PagingMode from app.rs into lib.rs
